+++ src/tacky/mod.rs
@@ -1,2 +1,3 @@
+pub mod display;
 pub mod from;
 pub mod program;
